+++ Projects/Feature/Sent/Sources/Envelope/EnvelopeProperty.swift
@@ -12,16 +12,16 @@ import Foundation
 
 struct EnvelopeProperty: Equatable, Hashable, Identifiable {
   /// 봉투의 친구 아이디 입니다.
-  var id: Int
+  var id: Int64
 
   /// envelopeUserName입니다.
   private let envelopeTargetUserName: String
   /// 전체 금액입니다.
-  private let totalPrice: Int
+  private let totalPrice: Int64
   /// 보낸 금액 입니다.
-  let totalSentPrice: Int
+  let totalSentPrice: Int64
   /// 받은 금액입니다.
-  let totalReceivedPrice: Int
+  let totalReceivedPrice: Int64
   /// envelopeUserName을 보여줍니다.
   var envelopeTargetUserNameText: String { envelopeTargetUserName }
   /// 전체 금액에 관한 Text입니다.
@@ -29,7 +29,7 @@ struct EnvelopeProperty: Equatable, Hashable, Identifiable {
     "전체 \(CustomNumberFormatter.formattedByThreeZero(totalPrice) ?? "")원"
   }
 
-  var receivedSubSentValue: Int {
+  var receivedSubSentValue: Int64 {
     return totalSentPrice - totalReceivedPrice
   }
 
@@ -37,11 +37,11 @@ struct EnvelopeProperty: Equatable, Hashable, Identifiable {
   var envelopeContents: [EnvelopeContent] = []
 
   init(
-    id: Int,
+    id: Int64,
     envelopeTargetUserName: String,
-    totalPrice: Int,
-    totalSentPrice: Int,
-    totalReceivedPrice: Int
+    totalPrice: Int64,
+    totalSentPrice: Int64,
+    totalReceivedPrice: Int64
   ) {
     self.id = id
     self.envelopeTargetUserName = envelopeTargetUserName
@@ -55,7 +55,7 @@ struct EnvelopeProperty: Equatable, Hashable, Identifiable {
 
 struct EnvelopeContent: Equatable, Hashable, Identifiable {
   /// 봉투의 아이디 입니다.
-  let id: Int
+  let id: Int64
   /// 봉투에 표시될 날짜 입ㄴ디ㅏ.
   let dateText: String
   /// 봉투의 Event Name 입니다.
@@ -63,7 +63,7 @@ struct EnvelopeContent: Equatable, Hashable, Identifiable {
   /// 봉투의 타입 입니다.
   let envelopeType: EnvelopeType
   /// 봉투의 가격 입니다.
-  let price: Int
+  let price: Int64
 
   var priceText: String {
     return CustomNumberFormatter.formattedByThreeZero(price) ?? "0원"
@@ -80,7 +80,7 @@ struct EnvelopeContent: Equatable, Hashable, Identifiable {
   ///   - eventName: 이벤트 이름을 입력합니다.
   ///   - envelopeType: 받았는지 보냈는지 enum 을통해 입력받습니다.
   ///   - price: 가격을 입력합니다.
-  init(id: Int, dateText: String, eventName: String, envelopeType: EnvelopeType, price: Int) {
+  init(id: Int64, dateText: String, eventName: String, envelopeType: EnvelopeType, price: Int64) {
     self.id = id
     self.dateText = dateText
     self.eventName = eventName
